Tests/LibWeb/Text/input/css/attr-serialization.html

<script src="../include.js"></script>
<script>
    test(() => {
        function serialize(input) {
            document.body.style.content = input;
            println(document.body.style.content);
        }

        serialize('attr(foo)');
        // FIXME: This should produce `attr(foo)` but doesn't yet.
        serialize('attr(    foo    )');
        serialize('attr(foo, "fallback")');
        // FIXME: This should produce `attr(foo, "fallback")` but doesn't yet.
        serialize('attr(   foo   ,    "fallback"   )');
        serialize('attr(foo string)');
        // FIXME: This should produce `attr(foo string)` but doesn't yet.
        serialize('attr(  foo   string  )');
        serialize('attr(foo string, "fallback")');
        // FIXME: This should produce `attr(foo string, "fallback")` but doesn't yet.
        serialize('attr(  foo   string  ,   "fallback"  )');
        serialize('  attr(foo)     attr(bar) attr(baz)  ');
    });
</script>
