Tests/LibWeb/Text/expected/css/attr-serialization.txt

attr(foo)
attr( foo )
attr(foo, "fallback")
attr( foo , "fallback" )
attr(foo string)
attr( foo string )
attr(foo string, "fallback")
attr( foo string , "fallback" )
attr(foo) attr(bar) attr(baz)
